Expert mode, checkboxes
* expert mode, field 'expert' in GlobalPage
* handling of checkboxes
* Util.readFileAsList()
* fix: using <xml> marker in configuration files
* fix: config values: fallback English if value not in selected language available
* improvements of test coverage

+++ pyunit/sqlitedbtest.py
@@ -32,7 +32,8 @@ class TestSqListeDb(unittest.TestCase):
 
 
     def test01Basic(self):
-        os.unlink(self._dbName)
+        if os.path.exists(self._dbName): 
+            os.unlink(self._dbName)
         self.setUp()
         self.assertTrue(None != self._db.getCursor())
